Final Version
Final version of the code.  The main helps generate a random set of values to simulate a patient during surgery.

Details
The output of an EKG is a repeating wave pattern which is divided into multiple segments. If the ST segment on an EKG varies too wildly, it can be indicative of ischemia.

You compute a running two minute average; the average is updated with new data every two seconds and consists of data for the preceding two minute period. In addition, you keep track of the minimum and maximum of each average. You want to sound an alarm if any of the following occur for any measurement:

1.	The minimum falls below -2
2.	The maximum falls above +2
3.	The difference (maximum-minimum) is greater than 2

In your program, the function ProcessEKG is automatically called every 2 seconds and passed the EKG object measurement. Your task is the write a routine simulation:

Where the EKG object contains the time (measured in seconds from 1/1/2000) and the value of the ST segment. If the conditions above are met, you want to call the Alarm function. ProcessEKG sits inside a class, and you can create global properties and functions in that class as needed.